Retool module for external mysql service and users
This commit strips out the current DB management and relies instead on
puppetlabs-mysql. It instead will restore a fresh dump of the mysql
database to start off with a blank wordpress site and adds a defined
resource type for creating additional wordpress admin users.

This rewrite is mostly for my own demonstration purposes but may
evolve into a reusable module with optional support for ssl.